+++ particula/backend/taichi/particles/properties/ti_kelvin_effect_module.py
@@ -10,13 +10,13 @@ _EXP_MAX = np.log(np.finfo(np.float64).max)        # ≈ 7.097e2
 
 # 3 – element-wise taichi funcs
 @ti.func
-def fget_kelvin_radius(σ: ti.f64, ρ: ti.f64, M: ti.f64, T: ti.f64) -> ti.f64:
+def fget_kelvin_radius(σ: float, ρ: float, M: float, T: float) -> float:
     R = ti.static(GAS_CONSTANT)
     return (2.0 * σ * M) / (R * T * ρ)
 
 
 @ti.func
-def fget_kelvin_term(r_p: ti.f64, r_k: ti.f64) -> ti.f64:
+def fget_kelvin_term(r_p: float, r_k: float) -> float:
     expo = r_k / r_p
     max_exp = ti.static(_EXP_MAX)
     expo = ti.min(expo, max_exp)    # overflow protection
@@ -25,11 +25,11 @@ def fget_kelvin_term(r_p: ti.f64, r_k: ti.f64) -> ti.f64:
 # 4 – kernels
 @ti.kernel
 def kget_kelvin_radius(
-    σ: ti.types.ndarray(dtype=ti.f64, ndim=1),
-    ρ: ti.types.ndarray(dtype=ti.f64, ndim=1),
-    M: ti.types.ndarray(dtype=ti.f64, ndim=1),
-    T: ti.f64,
-    res: ti.types.ndarray(dtype=ti.f64, ndim=1),
+    σ: ti.types.ndarray(dtype=float, ndim=1),
+    ρ: ti.types.ndarray(dtype=float, ndim=1),
+    M: ti.types.ndarray(dtype=float, ndim=1),
+    T: float,
+    res: ti.types.ndarray(dtype=float, ndim=1),
 ):
     for i in range(res.shape[0]):
         res[i] = fget_kelvin_radius(σ[i], ρ[i], M[i], T)
@@ -37,9 +37,9 @@ def kget_kelvin_radius(
 
 @ti.kernel
 def kget_kelvin_term(
-    r_p: ti.types.ndarray(dtype=ti.f64, ndim=2),
-    r_k: ti.types.ndarray(dtype=ti.f64, ndim=2),
-    res: ti.types.ndarray(dtype=ti.f64, ndim=2),
+    r_p: ti.types.ndarray(dtype=float, ndim=2),
+    r_k: ti.types.ndarray(dtype=float, ndim=2),
+    res: ti.types.ndarray(dtype=float, ndim=2),
 ):
     for I in ti.grouped(res):
         res[I] = fget_kelvin_term(r_p[I], r_k[I])
@@ -54,8 +54,8 @@ def ti_get_kelvin_radius(surface_tension, density, molar_mass, temperature):
 
     σ, ρ, M = map(np.atleast_1d, (surface_tension, density, molar_mass))
     n = σ.size
-    σ_ti, ρ_ti, M_ti = (ti.ndarray(dtype=ti.f64, shape=n) for _ in range(3))
-    res_ti = ti.ndarray(dtype=ti.f64, shape=n)
+    σ_ti, ρ_ti, M_ti = (ti.ndarray(dtype=float, shape=n) for _ in range(3))
+    res_ti = ti.ndarray(dtype=float, shape=n)
     σ_ti.from_numpy(σ)
     ρ_ti.from_numpy(ρ)
     M_ti.from_numpy(M)
@@ -85,9 +85,9 @@ def ti_get_kelvin_term(particle_radius, kelvin_radius_value):
         kr_mat = np.broadcast_to(kr, pr_mat.shape)    # (n, 1)
 
     shape_2d = pr_mat.shape                           # (rows, cols)
-    pr_ti = ti.ndarray(ti.f64, shape_2d);  pr_ti.from_numpy(pr_mat)
-    kr_ti = ti.ndarray(ti.f64, shape_2d);  kr_ti.from_numpy(kr_mat)
-    res_ti = ti.ndarray(ti.f64, shape_2d)
+    pr_ti = ti.ndarray(float, shape_2d);  pr_ti.from_numpy(pr_mat)
+    kr_ti = ti.ndarray(float, shape_2d);  kr_ti.from_numpy(kr_mat)
+    res_ti = ti.ndarray(float, shape_2d)
 
     kget_kelvin_term(pr_ti, kr_ti, res_ti)
     out = res_ti.to_numpy()
